requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
requests>=2.31.0
google-api-python-client>=2.129.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
